requirements.txt
fontmake
fonttools
ufolib2
pillow
watchdog
